提交 商品 推荐完成 待实现点击 以及滑动 联动

+++ app/src/main/java/com/example/juanshichang/bean/GridItemBean.kt
@@ -12,7 +12,7 @@ class GridItemBean {
     )
 
     data class Channel(
-        var channel_id: Int = 0,
+        var channel_id: Long = 0,
         var image_url: String = "",
         var order: Int = 0,
         var type: String = ""
